Cambios en Front y en Back ante revision del proyecto

+++ Hoteles-Back-End/src/user/user.controller.js
@@ -62,9 +62,11 @@ exports.login = async (req, res) => {
 //guardar
 exports.addUser = async (req, res) => {
     try {
-        /* let existUser = await User.findOne({ name: data.name })
-       if (existUser) return res.send({ message: 'Usuario ya creado' }) */
         let data = req.body
+
+        let existUser = await User.findOne({ name: data.name })
+        if (existUser) return res.status(400).send({ message: 'Usuario ya creado' })
+
         let params = {
             password: data.password
         }
@@ -103,7 +105,12 @@ exports.updateUser = async (req, res) => {
     try {
         let userId = req.params.id;
         let data = req.body
-        
+
+        let userDlete = await User.findOne({ _id: userId })
+        if (userDlete.role == "OWNER") {
+            return res.status(403).send({ message: 'No se puede editar al dueño del programa' })
+        }
+
         /* let token = req.user.sub;
         if (userId != token) return res.status(500).send({ message: "No tienes permiso para realizar esta accion" }) */
         if (data.password || Object.entries(data).length === 0 || data.role) return res.status(400).send({ message: 'Have submitted some data that cannot be updated' });
@@ -119,6 +126,28 @@ exports.updateUser = async (req, res) => {
     }
 }
 
+//update por token
+exports.updateToken = async (req, res) => {
+    try {
+        let userToken = req.user.sub
+        let data = req.body
+
+        if (data.password || Object.entries(data).length === 0 || data.role) return res.status(400).send({ message: 'Have submitted some data that cannot be updated' });
+        let userUp = await User.findOneAndUpdate(
+            { _id: userToken },
+            data,
+            { new: true }
+        )
+        if (!userUp) return res.status(404).send({ message: 'User not found and not updated' });
+        return res.send({ message: 'User updated', userUp })
+
+    } catch (err) {
+        console.log(err)
+        return res.status(500).send({ message: " Error getting users", error: err })
+    }
+}
+
+
 //delete
 exports.deleteUser = async (req, res) => {
     try {
@@ -126,6 +155,13 @@ exports.deleteUser = async (req, res) => {
         /* let token = req.user.sub;
         if(userId != token) return res.status(500).send({message: "No tienes permiso para realizar esta accion"}) */
         //eliminar el usuario
+
+        let userDlete = await User.findOne({ _id: userId })
+        if (userDlete.role == "OWNER") {
+            return res.status(403).send({ message: 'No se puede eliminar al dueño del programa' })
+        }
+
+
         let userDelete = await User.findByIdAndDelete({ _id: userId })
         if (!userDelete) return res.send({ message: "la cuenta no fue encontrado y por ende no eliminada" })
         return res.send({ message: `Cuenta con username ${userDelete.username} fue eliminada satisfactoriamente` })
@@ -135,23 +171,23 @@ exports.deleteUser = async (req, res) => {
     }
 }
 
-exports.getUsers = async(req, res)=>{
+exports.getUsers = async (req, res) => {
     try {
         let users = await User.find();
-        return res.send({message: 'User found', users})
+        return res.send({ message: 'User found', users })
     } catch (err) {
         console.error(err)
         return res.status(500).send({ message: " Error getting users" })
     }
 }
 
-exports.getUserId = async (req, res)=>{
+exports.getUserId = async (req, res) => {
     try {
-        let userId = req.params.id;
-        let user = await User.findOne({_id: userId})
-        if(!user) return res.status(404).send({message: 'User Not Found'})
-        return res.status(200).send({user})
-    } catch (err) { 
+        let userToken = req.user.sub
+        let user = await User.findOne({ _id: userToken })
+        if (!user) return res.status(404).send({ message: 'User Not Found' })
+        return res.status(200).send({ user })
+    } catch (err) {
         console.log(err)
     }
 }
